+++ Libplanet.Explorer.Tests/Queries/StateQueryTest.cs
@@ -14,6 +14,7 @@ using Libplanet.Blocks;
 using Libplanet.Consensus;
 using Libplanet.Crypto;
 using Libplanet.Explorer.Queries;
+using Libplanet.Store.Trie;
 using Xunit;
 using static Libplanet.Explorer.Tests.GraphQLTestUtils;
 
@@ -211,5 +212,7 @@ public class StateQueryTest
                         "032038e153d344773986c039ba5dbff12ae70cfdf6ea8beb7c5ea9b361a72a9233"),
                     new BigInteger(1)),
             });
+
+        public ITrie GetTrie(BlockHash offset) => null;
     }
 }