+++ Libplanet.Tests/Fixtures/IntegerSet.cs
@@ -87,6 +87,14 @@ namespace Libplanet.Tests.Fixtures
                 Store,
                 StateStore,
                 Genesis,
+                new ActionEvaluator(
+                    _ => policy.BlockAction,
+                    blockChainStates: new BlockChainStates(Store, StateStore),
+                    genesisHash: Genesis.Hash,
+                    nativeTokenPredicate: policy.NativeTokens.Contains,
+                    actionTypeLoader: StaticActionLoader.Create<Arithmetic>(),
+                    feeCalculator: null
+                ),
                 renderers: renderers ?? new[] { new ValidatingActionRenderer<Arithmetic>() }
             );
         }