+++ CHANGES.md
@@ -21,6 +21,11 @@ To be released.
        `RenderActionError(IAction, IActionContext, Exception)`
         method was changed to
         `RenderActionError(IValue, IActionContext, Exception)`.
+ -  Removed `trieGetter` parameter from `ActionEvaluator` constructor.
+    [[#3149]]
+ -  Added `IBlockChainStates.GetTrie()` interface method.  [[#3149]]
+ -  `BlockChain<T>.Create()` static method now requires `actionEvaluator`
+    parameter explicitly.  [[#3149]]
 
 ### Backward-incompatible network protocol changes
 
@@ -41,6 +46,7 @@ To be released.
 
 [#3117]: https://github.com/planetarium/libplanet/pull/3117
 [#3122]: https://github.com/planetarium/libplanet/pull/3122
+[#3149]: https://github.com/planetarium/libplanet/pull/3149
 
 
 Version 1.1.0